+++ src/core/interpreter/stepables.js
@@ -29,6 +29,7 @@ export {
     CallCCNode,
     BeginNode,
     ImportNode,
+    DefineLibraryNode,
     DynamicWindInit,
     RestoreContinuation,
     CallWithValuesNode,